change sidebar header buttons position
