+++ bigbluebutton-html5/imports/ui/components/chat/chat-dropdown/component.jsx
@@ -141,6 +141,7 @@ class ChatDropdown extends PureComponent {
               label={intl.formatMessage(intlMessages.options)}
               aria-label={intl.formatMessage(intlMessages.options)}
               onClick={() => null}
+              hideLabel
             />                    
           }
           opts={{
