+++ bigbluebutton-html5/imports/ui/components/common/control-header/styles.js
@@ -1,5 +1,7 @@
 import styled from 'styled-components';
 import { jumboPaddingY } from '/imports/ui/stylesheets/styled-components/general';
+import { fontSizeBase, headingsFontWeight} from '/imports/ui/stylesheets/styled-components/typography';
+import { colorGray } from '/imports/ui/stylesheets/styled-components/palette';
 
 const Header = styled.header`
   display: flex;
@@ -8,7 +10,14 @@ const Header = styled.header`
   padding-bottom: ${jumboPaddingY};
 `;
 
-const RightWrapper = styled.div`
+const Title = styled.h1`
+  font-size: ${fontSizeBase};
+  color: ${colorGray};
+  font-weight: ${headingsFontWeight};
+  margin: 0;
+`;
+
+const LeftWrapper = styled.div`
   & > div {
     display: flex;
   }
@@ -16,5 +25,6 @@ const RightWrapper = styled.div`
 
 export default {
   Header,
-  RightWrapper,
+  Title,
+  LeftWrapper,
 };
